% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildcust.R
\name{buildCust}
\alias{buildCust}
\title{Build Unique Customer Identifier}
\usage{
buildCust(numOfCust)
}
\arguments{
\item{numOfCust}{A number.}
}
\value{
A character with a unique customer identifiers
}
\description{
Builds a customer identifier. This is often used as a primary key of the customer dim table in databases.
}
\examples{
df <- buildCust(numOfCust = 1000)
df <- buildCust(numOfCust = 223)
}
