% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDistr.R
\name{buildDistr}
\alias{buildDistr}
\title{Build Data Distribution}
\usage{
buildDistr(st, en, cycles, trend)
}
\arguments{
\item{st}{A number}

\item{en}{A number.}

\item{cycles}{A string.}

\item{trend}{A number.}
}
\value{
A dataframe with data distribution is returned.
}
\description{
Build Data Distribution
}
