% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModelData.R
\name{buildModelData}
\alias{buildModelData}
\title{Generate Synthetic Data using Uncovr API}
\usage{
buildModelData(numOfObs, numOfVars, key)
}
\arguments{
\item{numOfObs}{A number. This represents the number of observations in the data. In other words, the number of rows of data that are requested to be generated. The \emph{numOfObs} argument must be a non-negative integer and in the current version, this function accepts a range of 100 to 10,000.}

\item{numOfVars}{A number. This represents the number of independent variables in the data. In other words, the number of columns besides the dependent variable of data that are requested to be generated. The \emph{numOfVars} argument must be a non-negative integer and in the current version, this function accepts a range of 1 to 100.}

\item{key}{An alpha numeric. This is the subscription key that can be sourced from the developer portal of uncovr API available at \emph{https://foyi.developer.azure-api.net/}.}
}
\value{
A .
}
\description{
Generate Synthetic Data using Uncovr API
}
