% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildName.R
\name{buildName}
\alias{buildName}
\title{Build Dynamic Strings}
\usage{
buildName(numOfItems, prefix)
}
\arguments{
\item{numOfItems}{A number.}

\item{prefix}{A string.}
}
\value{
A character with the strings is returned. These strings use the prefix that is mentioned in the argument "prefix"
}
\description{
Builds strings that could be further used as identifiers.This is an internal function and is currently not exported in the package.
}
