% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildOutliers.R
\name{buildOutliers}
\alias{buildOutliers}
\title{Build Outliers in Data Distribution}
\usage{
buildOutliers(distr)
}
\arguments{
\item{distr}{numeric vector}
}
\value{
A numeric vector with random values replaced with outlier values
}
\description{
Builds outlier values and replaces random data points with this outliers. This is an internal function and is currently not exported in the package.
}
