% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPareto.R
\name{buildPareto}
\alias{buildPareto}
\title{Map Factors Based on Pareto Arguments}
\usage{
buildPareto(factor1, factor2, pareto)
}
\arguments{
\item{factor1}{A factor.}

\item{factor2}{A factor.}

\item{pareto}{A vector of 2 numbers.}
}
\value{
A dataframe with factor 1 and factor 2 as columns. Based on the pareto arguments passed, the column factor 1 is mapped to factor 2.
}
\description{
Map Factors Based on Pareto Arguments
}
