% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSpike.R
\name{buildSpike}
\alias{buildSpike}
\title{Build Spikes in the Data Distribution}
\usage{
buildSpike(distr, spike)
}
\arguments{
\item{distr}{numeric vector. This is the input vector for which the spike value needs to be set.}

\item{spike}{A number. This represents the seasonality of data. It can take any value from 1 to 12. These numbers represent months in a year, from January to December respectively. For example, if the spike is set to 12, it means that December has the highest number of transactions. This is an internal function and is currently not exported in the package.}
}
\value{
A numeric vector reordered
}
\description{
Builds spikes in the data distribution. For example, in retail industry transactions are generally higher during the holiday season such as December. This function is used to set the same.
}
