% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genFirstPairs.R
\name{genFirstPairs}
\alias{genFirstPairs}
\title{Extracts the First Two Alphabets of the String}
\usage{
genFirstPairs(s)
}
\arguments{
\item{s}{A string. This is the string from which the first two alphabets are to be extracted.}
}
\value{
First two alphabets of the string input.
}
\description{
For a given string, this function extracts the first two alphabets. This function is further used by \code{\link{genMatrix}} function.
}
