% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTriples.R
\name{genTriples}
\alias{genTriples}
\title{Extracts Three Consecutive Alphabets of the String}
\usage{
genTriples(s)
}
\arguments{
\item{s}{A string. This is the string from which three consecutive alphabets are to be extracted.}
}
\value{
List of three alphabet combinations of the string input.
}
\description{
For a given string, this function extracts three consecutive alphabets. This function is further used by \code{\link{genMatrix}} function.
}
