% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingArgHandler.R
\name{missingArgHandler}
\alias{missingArgHandler}
\title{Handle Missing Arguments in Function}
\usage{
missingArgHandler(argMissed, argDefault)
}
\arguments{
\item{argMissed}{This is the argument that needs to be handled.}

\item{argDefault}{This is the default value of the argument that is missing in the function called.}
}
\value{
The default value of the missing argument.
}
\description{
Replaces the missing argument with the default value. This is an internal function and is currently not exported in the package.
}
\details{
This function plays the role of error handler by setting the default values of the arguments when a function is called without specifying any arguments.
}
