% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeDf.R
\name{treeDf}
\alias{treeDf}
\title{A supporting function.}
\usage{
treeDf(...)
}
\arguments{
\item{...}{This is a placeholder argument.}
}
\value{
A dataframe.
}
\description{
This is used by another internal function \code{\link{genTree}}.
}
