% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncovrApi.R
\name{uncovrApi}
\alias{uncovrApi}
\title{POST function for calling uncovr API}
\usage{
uncovrApi(body, key)
}
\arguments{
\item{body}{A number.}

\item{key}{An alpha numeric.}
}
\value{
A .
#function to call uncovr api
}
\description{
POST function for calling uncovr API
}
