% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildcust.R
\name{buildCust}
\alias{buildCust}
\title{Build Unique Customer Identifier}
\usage{
buildCust(numOfCust)
}
\arguments{
\item{numOfCust}{A number.}
}
\value{
A character with a unique customer identifiers
}
\description{
Build Unique Customer Identifier
}
\examples{
df <- buildCust(1000)
df <- buildCust(223)
}
