% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildName.R
\name{buildName}
\alias{buildName}
\title{Build Dynamic Strings}
\usage{
buildName(numOfItems, prefix)
}
\arguments{
\item{numOfItems}{A number.}

\item{prefix}{A string.}
}
\value{
A character with the strings is returned. These strings use the prefix that is mentioned in the argument "prefix"
}
\description{
Build Dynamic Strings
}
