% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildName.R
\name{buildName}
\alias{buildName}
\title{Build Dynamic Strings}
\usage{
buildName(numOfItems, prefix)
}
\arguments{
\item{numOfItems}{A number. This defines the number of elements to be output.}

\item{prefix}{A string. This defines the prefix for the strings. For example, the function buildCust uses this function and passes the prefix "cust" while the function buildProd passes the prefix "sku"}
}
\value{
A character with the alphanumeric strings is returned. These strings use the prefix that is mentioned in the argument "prefix"
}
\description{
Builds strings that could be further used as identifiers. This is an internal function and is currently not exported in the package.
}
\details{
This function is used by other internal functions namely, buildCust and buildProd to produce the alphanumeric identifiers for customers and products respectively.
}
