% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildProd.R
\name{buildProd}
\alias{buildProd}
\title{Build Product Data}
\usage{
buildProd(numOfProd, minPrice, maxPrice)
}
\arguments{
\item{numOfProd}{A number.}

\item{minPrice}{A number.}

\item{maxPrice}{A number.}
}
\value{
A character with product identifier and price
}
\description{
Build Product Data
}
\examples{
df <- buildProd(1000, 5, 100)
df <- buildProd(29,3,50)
}
