% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildProd.R
\name{buildProd}
\alias{buildProd}
\title{Build Product Data}
\usage{
buildProd(numOfProd, minPrice, maxPrice)
}
\arguments{
\item{numOfProd}{A number. This defines the number of unique products.}

\item{minPrice}{A number. This defines the lower threshold of the product price.}

\item{maxPrice}{A number. This defines the lower threshold of the product price.}
}
\value{
A character with product identifier and price.
}
\description{
Builds unique product identifier and price. The price of the product is generated randomly within the minimum and the maximum range provided as input.
}
\examples{
df <- buildProd(numOfProd = 1000, minPrice = 5, maxPrice = 100)
df <- buildProd(numOfProd = 29, minPrice = 3, maxPrice = 50)
}
