% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSpike.R
\name{buildSpike}
\alias{buildSpike}
\title{Build Spikes in the Data Distribution}
\usage{
buildSpike(distr, spike)
}
\arguments{
\item{distr}{numeric.}

\item{spike}{A number. This defines the month where the data distribution has the highest value. The number ranges from 1 to 12 representing the 12 months in the order of January to December. This is an internal function and is currently not exported in the package.}
}
\value{
A numeric vector reordered
}
\description{
Builds spikes in the data distribution. For example, in retail industry transactions are generally higher during the holiday season such as December.
}
