% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factTransactions.R
\name{genTrans}
\alias{genTrans}
\title{Build Transaction Data}
\usage{
genTrans(cycles, trend, transactions)
}
\arguments{
\item{cycles}{A string.}

\item{trend}{A number.}

\item{transactions}{A number.}
}
\value{
A dataframe with day number and count of transactions on that day
}
\description{
Build Transaction Data
}
\examples{
df <- genTrans("y", 1,10000)
df <- genTrans("q", -1,32000)
}
