% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingArgHandler.R
\name{missingArgHandler}
\alias{missingArgHandler}
\title{Handle Missing Arguments in Function}
\usage{
missingArgHandler(argMissed, argDefault)
}
\arguments{
\item{argMissed}{A variable. This is the argument that needs to be handled.}

\item{argDefault}{A variable. This is the default value.}
}
\value{
The default value of the missing argument
}
\description{
Replaces the missing argument with the default value. This is an internal function and is currently not exported in the package.
}
