% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingArgHandler.R
\name{missingArgHandler}
\alias{missingArgHandler}
\title{Handle Missing Arguments in Function}
\usage{
missingArgHandler(argMissed, argDefault)
}
\arguments{
\item{argMissed}{A variable.}

\item{argDefault}{A variable.}
}
\value{
The default value of the missing argument
}
\description{
Handle Missing Arguments in Function
}
