% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncovrApi.R
\name{uncovrApi}
\alias{uncovrApi}
\title{POST Function for Calling uncovr API}
\usage{
uncovrApi(body, key)
}
\arguments{
\item{body}{A json with the details of the independent and dependent variable.}

\item{key}{An alpha numeric. This is the subscription key that can be sourced from the developer portal of uncovr API available at \emph{https://foyi.developer.azure-api.net/}.}
}
\value{
A json.
}
\description{
This function makes the POST call to the \emph{uncovr} API.
}
\details{
The purpose of this function can be best understood when explained within the context that is given below. There is a closed source SaaS(Software as a Service) software named \emph{uncovr} that provides an API(Application Programming Interface). In its current state, the SaaS software is free to use with some constraints around the volume of data and the frequency of API calls. One of the functions of \emph{uncovr} API takes an input of number of observations i.e. rows and number of independent variables namely columns and gives an output. This function \emph{uncovrApi} makes the connection to \emph{uncovr} API and sources the response.
#function to call uncovr api
}
